refactor(karrotframe): 위치 수정
